[refactor] rearrange directory
data와 presentation으로 폴더 구분
