+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".login.MyPageActivity"
+            android:name=".presentation.login.MyPageActivity"
             android:exported="false" />
         <activity
-            android:name=".login.SignUpActivity"
+            android:name=".presentation.login.SignUpActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".login.LoginActivity"
+            android:name=".presentation.login.LoginActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
@@ -30,7 +30,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".main.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false" />
     </application>
