+++ app/src/main/java/org/android/go/sopt/presentation/album/AlbumAdapter.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.album
+package org.android.go.sopt.presentation.album
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
