+++ app/src/main/java/org/android/go/sopt/presentation/album/AlbumFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.album
+package org.android.go.sopt.presentation.album
 
 import android.os.Bundle
 import android.view.LayoutInflater
