+++ app/src/main/java/org/android/go/sopt/presentation/follower/FollowerFragment.kt
@@ -1,15 +1,15 @@
-package org.android.go.sopt.search
+package org.android.go.sopt.presentation.follower
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.android.go.sopt.databinding.FragmentSearchBinding
+import org.android.go.sopt.databinding.FragmentFollowerBinding
 
-class SearchFragment : Fragment() {
-    private var _binding: FragmentSearchBinding? = null
-    private val binding: FragmentSearchBinding
+class FollowerFragment : Fragment() {
+    private var _binding: FragmentFollowerBinding? = null
+    private val binding: FragmentFollowerBinding
         get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
 
     override fun onCreateView(
@@ -17,7 +17,7 @@ class SearchFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSearchBinding.inflate(inflater, container, false)
+        _binding = FragmentFollowerBinding.inflate(inflater, container, false)
         return binding.root
     }
