+++ app/src/main/java/org/android/go/sopt/presentation/playlist/ListFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.presentation.playlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
