+++ app/src/main/java/org/android/go/sopt/presentation/playlist/PlaylistSong.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.presentation.playlist
 
 import androidx.annotation.DrawableRes
