+++ app/src/main/java/org/android/go/sopt/presentation/playlist/PlaylistTitle.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.presentation.playlist
 
 data class PlaylistTitle(
     val id: Int,
