+++ app/src/main/java/org/android/go/sopt/presentation/playlist/PlaylistTitleAdapter.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.presentation.playlist
 
 import android.content.Context
 import android.view.LayoutInflater
