+++ app/src/main/java/org/android/go/sopt/presentation/playlist/mockPlayListTitle.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.playlist
+package org.android.go.sopt.presentation.playlist
 
 val mockPlayListTitle = listOf<PlaylistTitle>(
     PlaylistTitle(0, "상호의 플레이리스트")
